+++ SylvaniaLightify.indigoPlugin/Contents/Info.plist
@@ -3,9 +3,9 @@
 <plist version="1.0">
 <dict>
 	<key>PluginVersion</key>
-	<string>1.0.5</string>
+	<string>2.0.0</string>
 	<key>ServerApiVersion</key>
-	<string>2.0</string>
+	<string>3.0</string>
 	<key>IwsApiVersion</key>
 	<string>1.0.0</string>
 	<key>CFBundleDisplayName</key>
