+++ SylvaniaLightify.indigoPlugin/Contents/Server Plugin/plugin.py
@@ -9,13 +9,13 @@
 # Imports
 ################################################################################
 import indigo
-import urllib2
+import urllib.request, urllib.error, urllib.parse
 from xml.dom.minidom import parseString
 import lightify
 import time
 import datetime
 import threading
-import Queue
+import queue
 import sys
 import enum
 import logging
@@ -47,7 +47,7 @@ class Plugin(indigo.PluginBase):
         self.deviceList = []
         self.deviceThreads = []
         self.lightifyLock = threading.Lock()
-        self.lightifyQueue = Queue.Queue(100)
+        self.lightifyQueue = queue.Queue(100)
         self.lastRefreshTime = datetime.datetime.now()
         self.pluginVersion = pluginVersion
 
@@ -93,8 +93,8 @@ class Plugin(indigo.PluginBase):
     # Startup
     ########################################
     def startup(self):
-        indigo.server.log(u"Startup - SylvaniaLightify Plugin, version=" + self.pluginVersion)
-        indigo.server.log(u"Initializing Lightify Hub, IP Address=" + self.lightifyHubIpAddr)
+        indigo.server.log("Startup - SylvaniaLightify Plugin, version=" + self.pluginVersion)
+        indigo.server.log("Initializing Lightify Hub, IP Address=" + self.lightifyHubIpAddr)
         try:
             loglevel = logging.INFO
             if self.debug:
@@ -121,7 +121,7 @@ class Plugin(indigo.PluginBase):
             scenes = list()
 
         if not self.pluginPrefs.get('scenes', False) and not self.pluginPrefs.get('scenesList', False):
-            indigo.server.log(u"pluginPrefs lacks scenes.  Adding.")
+            indigo.server.log("pluginPrefs lacks scenes.  Adding.")
             # Add the empty scenes list to the prefs.
             for aNumber in range(1,21):
                 # Create a blank sub-list for storing scene name and scene states.
@@ -130,7 +130,7 @@ class Plugin(indigo.PluginBase):
                 scenes.append(scene)
 
                 if aNumber == 1:
-                    self.debugLog(u"pluginPrefs lacks scenes.  Adding Christmas.")
+                    self.debugLog("pluginPrefs lacks scenes.  Adding Christmas.")
                     # Add the scene name.
                     scene.append('Christmas')
                     scene.append('rotate-colors')
@@ -147,7 +147,7 @@ class Plugin(indigo.PluginBase):
                     circadianData.append('10,35,90,95,100,80,40')
                     scene.append(circadianData)
                 elif aNumber == 2:
-                    self.debugLog(u"pluginPrefs lacks scenes.  Adding Halloween.")
+                    self.debugLog("pluginPrefs lacks scenes.  Adding Halloween.")
                     # Add the scene name.
                     scene.append('Halloween')
                     scene.append('match-colors')
@@ -163,7 +163,7 @@ class Plugin(indigo.PluginBase):
                     circadianData.append('10,35,90,95,100,80,40')
                     scene.append(circadianData)
                 elif aNumber == 3:
-                    self.debugLog(u"pluginPrefs lacks scenes.  Adding St. Patricks.")
+                    self.debugLog("pluginPrefs lacks scenes.  Adding St. Patricks.")
                     scene.append('St. Patricks')
                     scene.append('match-colors')
                     scene.append('60')
@@ -178,7 +178,7 @@ class Plugin(indigo.PluginBase):
                     circadianData.append('10,35,90,95,100,80,40')
                     scene.append(circadianData)
                 elif aNumber == 4:
-                    self.debugLog(u"pluginPrefs lacks scenes.  Adding IndoorCirc.")
+                    self.debugLog("pluginPrefs lacks scenes.  Adding IndoorCirc.")
                     scene.append('IndoorCirc')
                     scene.append('circadian')
                     scene.append('600')
@@ -191,7 +191,7 @@ class Plugin(indigo.PluginBase):
                     circadianData.append('10,35,90,95,100,80,40')
                     scene.append(circadianData)
                 elif aNumber == 5:
-                    self.debugLog(u"pluginPrefs lacks scenes.  Adding OutdoorCirc.")
+                    self.debugLog("pluginPrefs lacks scenes.  Adding OutdoorCirc.")
                     scene.append('OutdoorCirc')
                     scene.append('circadian')
                     scene.append('600')
@@ -204,7 +204,7 @@ class Plugin(indigo.PluginBase):
                     circadianData.append('20,55,80,95,100,85,60')
                     scene.append(circadianData)
                 elif aNumber == 6:
-                    self.debugLog(u"pluginPrefs lacks scenes.  Adding MBR Circ.")
+                    self.debugLog("pluginPrefs lacks scenes.  Adding MBR Circ.")
                     scene.append('MBR Circ')
                     scene.append('circadian')
                     scene.append('600')
@@ -218,17 +218,17 @@ class Plugin(indigo.PluginBase):
                     scene.append(circadianData)
                 else:
                     # Add the scene name.
-                    scene.append('Scene ' + unicode(aNumber))
+                    scene.append('Scene ' + str(aNumber))
                     scene.append('match-colors')
                     scene.append('60')
 
             # Add the new list of empty scenes to the prefs.
             self.pluginPrefs['scenes'] = scenes
-            self.debugLog(u"pluginPrefs now contains 20 scenes.")
+            self.debugLog("pluginPrefs now contains 20 scenes.")
 
         # If scenes exist, make sure there are 20 of them.
         if not self.pluginPrefs.get('scenesList', False) or len(self.pluginPrefs['scenesList']) == 0:
-            indigo.server.log(u"pluginPrefs lacks scenesList.  Adding.")
+            indigo.server.log("pluginPrefs lacks scenesList.  Adding.")
             # fill in the dictionary instead of the list
             self.pluginPrefs['scenesList'] = list()
             # Start a new list of empty scenes.
@@ -266,14 +266,14 @@ class Plugin(indigo.PluginBase):
                         circadianDict['CircadianColorTempValues'] = '1900,2200,5500,6400,6500,4500,2200'
                         circadianDict['CircadianBrightnessValues'] = '10,35,90,98,100,80,35'
 
-                self.debugLog(u"adding to scenesList. Scene name: " + str(sceneDict['sceneName']))
+                self.debugLog("adding to scenesList. Scene name: " + str(sceneDict['sceneName']))
 
             # Add the new list of empty scenes to the prefs.
             self.pluginPrefs['scenesList'] = scenesList
 
         # after creating the scenesList, we can remove the old scenes
         if 'scenes' in self.pluginPrefs:
-            self.debugLog(u"removing old 'scenes' from pluginPrefs.")
+            self.debugLog("removing old 'scenes' from pluginPrefs.")
             del self.pluginPrefs['scenes']
 
 
@@ -295,7 +295,7 @@ class Plugin(indigo.PluginBase):
                             self.debugLog('...Refreshing Lightify Device Status')
                             self.lightifyConn.update_all_light_status()
                             self.lightifyConn.update_group_list()
-                            for (addr, light) in self.lightifyConn.lights().iteritems():
+                            for (addr, light) in self.lightifyConn.lights().items():
                                 self.debugLog('...light details=' + str(light) + ', on=' + str(light.on()))
 
                             # now we cycle through each group
@@ -427,7 +427,7 @@ class Plugin(indigo.PluginBase):
                                     self.debugLog('...runConcurrentThread RELEASING lightifyLock')
                                     self.lightifyLock.release()
                         self.debugLog('...runConcurrentThread Done, num_tries=' + str(num_tries))
-                except Queue.Empty:
+                except queue.Empty:
                     indigo.server.log('...runConcurrentThread EXCEPTION EMPTY QUEUE, EMPTY QUEUE, EMPTY QUEUE, EMPTY QUEUE')
 
         except self.StopThread:
@@ -436,9 +436,9 @@ class Plugin(indigo.PluginBase):
 ########################################
     def groupListGenerator(self, filter="", valuesDict=None, typeId="", targetId=0):
         # Used in actions that need a list of Lightify groups.
-        self.debugLog(u"Starting groupListGenerator.\n  filter: "
-                      + unicode(filter) + u"\n  valuesDict: " + unicode(valuesDict) + u"\n  typeId: "
-                      + unicode(typeId) + u"\n  targetId: " + unicode(targetId))
+        self.debugLog("Starting groupListGenerator.\n  filter: "
+                      + str(filter) + "\n  valuesDict: " + str(valuesDict) + "\n  typeId: "
+                      + str(typeId) + "\n  targetId: " + str(targetId))
 
         returnGroupList = list()
 
@@ -449,7 +449,7 @@ class Plugin(indigo.PluginBase):
             groupId = groupId + 1
 
         # Debug
-        self.debugLog(u"groupListGenerator: Return bulb list is %s" % returnGroupList)
+        self.debugLog("groupListGenerator: Return bulb list is %s" % returnGroupList)
 
         return returnGroupList
 
@@ -458,14 +458,14 @@ class Plugin(indigo.PluginBase):
     ########################################
     def sceneListGenerator(self, filter="", valuesDict=None, typeId="", deviceId=0):
         # Used by action dialogs to generate a list of Scenes saved in the plugin prefs.
-        self.debugLog(u"Starting sceneListGenerator.\n  filter: " + unicode(filter) + u"\n  valuesDict: "
-                      + unicode(valuesDict) + u"\n  typeId: " + unicode(typeId) + u"\n  deviceId: "
-                      + unicode(deviceId))
+        self.debugLog("Starting sceneListGenerator.\n  filter: " + str(filter) + "\n  valuesDict: "
+                      + str(valuesDict) + "\n  typeId: " + str(typeId) + "\n  deviceId: "
+                      + str(deviceId))
 
         theList = list()	# Menu item list.
 
         scenesList = self.pluginPrefs.get('scenesList', None)
-        self.debugLog(u"sceneListGenerator: Scenes in plugin prefs:\n" + unicode(scenesList))
+        self.debugLog("sceneListGenerator: Scenes in plugin prefs:\n" + str(scenesList))
 
         if scenesList != None:
             sceneNumber = 0
@@ -475,12 +475,12 @@ class Plugin(indigo.PluginBase):
                 hasData = ""
                 sceneName = sceneDict['sceneName']
                 self.debugLog('scene ' + sceneName)
-                if 'colorData' in sceneDict.keys() or 'circadianData' in sceneDict.keys():
+                if 'colorData' in list(sceneDict.keys()) or 'circadianData' in list(sceneDict.keys()):
                     self.debugLog("sceneListGenerator: scene " + sceneName + ", hasData=True")
                     hasData = "*"
 
                 sceneNumber += 1
-                theList.append((sceneNumber, hasData + unicode(sceneNumber) + ": " + sceneName))
+                theList.append((sceneNumber, hasData + str(sceneNumber) + ": " + sceneName))
         else:
             theList.append((0, "-- no scenes --"))
 
@@ -490,9 +490,9 @@ class Plugin(indigo.PluginBase):
     ########################################
     def activeSceneListGenerator(self, filter="", valuesDict=None, typeId="", deviceId=0):
         # Used by action dialogs to generate a list of Scenes saved in the plugin prefs.
-        self.debugLog(u"Starting activeSceneListGenerator.\n  filter: " + unicode(filter) + u"\n  valuesDict: "
-                      + unicode(valuesDict) + u"\n  typeId: " + unicode(typeId) + u"\n  deviceId: "
-                      + unicode(deviceId))
+        self.debugLog("Starting activeSceneListGenerator.\n  filter: " + str(filter) + "\n  valuesDict: "
+                      + str(valuesDict) + "\n  typeId: " + str(typeId) + "\n  deviceId: "
+                      + str(deviceId))
 
         theList = list()	# Menu item list.
 
@@ -504,13 +504,13 @@ class Plugin(indigo.PluginBase):
                 # Determine whether the Scene has saved data or not.
                 hasData = False
                 sceneName = sceneDict['sceneName']
-                if 'colorData' in sceneDict.keys() or 'circadianData' in sceneDict.keys():
+                if 'colorData' in list(sceneDict.keys()) or 'circadianData' in list(sceneDict.keys()):
                     self.debugLog("activeSceneListGenerator: scene " + sceneName + ", hasData=True")
                     hasData = True
 
                 sceneNumber += 1
                 if hasData is True:
-                    theList.append((sceneNumber, unicode(sceneNumber) + ": " + sceneName))
+                    theList.append((sceneNumber, str(sceneNumber) + ": " + sceneName))
         else:
             theList.append((0, "-- no scenes --"))
 
@@ -520,12 +520,12 @@ class Plugin(indigo.PluginBase):
     # Scenes List Item Selected (callback from action UI)
     ########################################
     def scenesListItemSelected(self, valuesDict=None, typeId="", deviceId=0):
-        self.debugLog(u"Starting scenesListItemSelected.  valuesDict: " + unicode(valuesDict)
-                      + u", typeId: " + unicode(typeId) + u", targetId: " + unicode(deviceId))
+        self.debugLog("Starting scenesListItemSelected.  valuesDict: " + str(valuesDict)
+                      + ", typeId: " + str(typeId) + ", targetId: " + str(deviceId))
 
         self.sceneListSelection = valuesDict['sceneId']
         sceneId = int(valuesDict['sceneId'])
-        self.debugLog(u".. sceneId-" + self.sceneListSelection)
+        self.debugLog(".. sceneId-" + self.sceneListSelection)
 
         sceneDict = self.pluginPrefs['scenesList'][sceneId-1]
 
@@ -533,49 +533,49 @@ class Plugin(indigo.PluginBase):
         valuesDict['sceneType'] = sceneDict['sceneType']
         valuesDict['sceneInterval'] = sceneDict['sceneInterval']
 
-        if 'colorData' in sceneDict.keys():
+        if 'colorData' in list(sceneDict.keys()):
             colorDict = sceneDict['colorData']
 
-            if 'setting1Value' in colorDict.keys():
+            if 'setting1Value' in list(colorDict.keys()):
                 valuesDict['setting1Value'] = colorDict['setting1Value']
             else:
                 valuesDict['setting1Value'] = "255,128,128,0,50,100"
-            if 'setting2Value' in colorDict.keys():
+            if 'setting2Value' in list(colorDict.keys()):
                 valuesDict['setting2Value'] = colorDict['setting2Value']
             else:
                 valuesDict['setting2Value'] = "128,255,255,0,75,100"
-            if 'setting3Value' in colorDict.keys():
+            if 'setting3Value' in list(colorDict.keys()):
                 valuesDict['setting3Value'] = colorDict['setting3Value']
             else:
                 valuesDict['setting3Value'] = ""
-            if 'setting4Value' in colorDict.keys():
+            if 'setting4Value' in list(colorDict.keys()):
                 valuesDict['setting4Value'] = colorDict['setting4Value']
             else:
                 valuesDict['setting4Value'] = ""
-            if 'setting5Value' in colorDict.keys():
+            if 'setting5Value' in list(colorDict.keys()):
                 valuesDict['setting5Value'] = colorDict['setting5Value']
             else:
                 valuesDict['setting5Value'] = ""
 
-        if 'circadianData' in sceneDict.keys():
+        if 'circadianData' in list(sceneDict.keys()):
             circDict = sceneDict['circadianData']
 
-            if 'CircadianColorTempValues' in circDict.keys():
+            if 'CircadianColorTempValues' in list(circDict.keys()):
                 valuesDict['CircadianColorTempValues'] = circDict['CircadianColorTempValues']
             else:
                 valuesDict['CircadianColorTempValues'] = "1900,2300,4000,6400,6500,4500,2200"
-            if 'CircadianBrightnessValues' in circDict.keys():
+            if 'CircadianBrightnessValues' in list(circDict.keys()):
                 valuesDict['CircadianBrightnessValues'] = circDict['CircadianBrightnessValues']
             else:
                 valuesDict['CircadianBrightnessValues'] = "15,35,80,95,100,80,40"
 
-        self.debugLog(u"... selected sceneId: " + unicode(sceneId)
-                      + u", sceneName: " + unicode(sceneDict['sceneName']) + u", sceneType: " + unicode(sceneDict['sceneType'])
-                      + u", sceneInterval: " + unicode(sceneDict['sceneInterval']))
-        if 'colorData' in sceneDict.keys():
-            self.debugLog(u"... colorData:+ u" + unicode(sceneDict['colorData']))
-        if 'circadianData' in sceneDict.keys():
-            self.debugLog(u"... circadianData:+ u" + unicode(sceneDict['circadianData']))
+        self.debugLog("... selected sceneId: " + str(sceneId)
+                      + ", sceneName: " + str(sceneDict['sceneName']) + ", sceneType: " + str(sceneDict['sceneType'])
+                      + ", sceneInterval: " + str(sceneDict['sceneInterval']))
+        if 'colorData' in list(sceneDict.keys()):
+            self.debugLog("... colorData:+ u" + str(sceneDict['colorData']))
+        if 'circadianData' in list(sceneDict.keys()):
+            self.debugLog("... circadianData:+ u" + str(sceneDict['circadianData']))
 
         return valuesDict
 
@@ -583,23 +583,23 @@ class Plugin(indigo.PluginBase):
     # Save Scene Action
     ########################################
     def saveScene(self, valuesDict, typeId):
-        self.debugLog(u"Starting saveScene. valuesDict values:\n" + unicode(valuesDict)
-                      +  ", typeId:\n" + unicode(typeId))
+        self.debugLog("Starting saveScene. valuesDict values:\n" + str(valuesDict)
+                      +  ", typeId:\n" + str(typeId))
         errorsDict = indigo.Dict()
-        errorsDict['showAlertText'] = u""
+        errorsDict['showAlertText'] = ""
 
         sceneId = valuesDict['sceneId']
         sceneName = valuesDict['sceneName']
         sceneType = valuesDict['sceneType']
         sceneInterval = valuesDict['sceneInterval']
 
-        self.debugLog(u"Starting saveScene. sceneId:" + unicode(sceneId)
-                      +  ", name:" + unicode(sceneName)
-                      +  ", interval:" + unicode(sceneInterval)
-                      +  ", type:" + unicode(sceneType))
+        self.debugLog("Starting saveScene. sceneId:" + str(sceneId)
+                      +  ", name:" + str(sceneName)
+                      +  ", interval:" + str(sceneInterval)
+                      +  ", type:" + str(sceneType))
 
         if not sceneId:
-            errorText = u"No Scene specified."
+            errorText = "No Scene specified."
             self.errorLog(errorText)
             # Remember the error.
             self.lastErrorMessage = errorText
@@ -610,15 +610,15 @@ class Plugin(indigo.PluginBase):
             # Subtract 1 because key values are 0-based.
             sceneId -= 1
 
-        self.debugLog(u"saveScene - starting Validation for Scene " + unicode(sceneId + 1) + u" (" + sceneName + u") .")
+        self.debugLog("saveScene - starting Validation for Scene " + str(sceneId + 1) + " (" + sceneName + ") .")
         returnVal = True
         if sceneType == 'match-colors' or sceneType == 'rotate-colors':
             # must have at least 1 color setting
             configStr = 'setting1Value'
             if not valuesDict[configStr]:
-                errorsDict[configStr] =  u"No " + configStr + u" specified."
+                errorsDict[configStr] =  "No " + configStr + " specified."
                 errorsDict['showAlertText'] += errorsDict[configStr]
-                errorText = u"No " + configStr + u" specified."
+                errorText = "No " + configStr + " specified."
                 self.errorLog(errorText)
                 returnVal = False
             else:
@@ -628,16 +628,16 @@ class Plugin(indigo.PluginBase):
             # must have at least 1 color setting
             configStr = 'CircadianColorTempValues'
             if not valuesDict[configStr]:
-                errorsDict[configStr] =  u"No " + configStr + u" specified."
+                errorsDict[configStr] =  "No " + configStr + " specified."
                 errorsDict['showAlertText'] += errorsDict[configStr]
-                errorText = u"No " + configStr + u" specified."
+                errorText = "No " + configStr + " specified."
                 self.errorLog(errorText)
                 returnVal = False
             configStr = 'CircadianBrightnessValues'
             if not valuesDict[configStr]:
-                errorsDict[configStr] =  u"No " + configStr + u" specified."
+                errorsDict[configStr] =  "No " + configStr + " specified."
                 errorsDict['showAlertText'] += errorsDict[configStr]
-                errorText = u"No " + configStr + u" specified."
+                errorText = "No " + configStr + " specified."
                 self.errorLog(errorText)
                 returnVal = False
 
@@ -647,11 +647,11 @@ class Plugin(indigo.PluginBase):
                 returnVal, valuesDict, errorsDict = self.validateCircadianConfig('CircadianBrightnessValues', valuesDict, errorsDict)
 
         if returnVal is False:
-            self.debugLog(u"saveScene - errors in validation Scene " + unicode(sceneId + 1) + u" (" + sceneName + u") ."
-                          + unicode(str(errorsDict['showAlertText'])))
+            self.debugLog("saveScene - errors in validation Scene " + str(sceneId + 1) + " (" + sceneName + ") ."
+                          + str(str(errorsDict['showAlertText'])))
             return valuesDict, errorsDict
 
-        self.debugLog(u"saveScene - finished Validation for Scene " + unicode(sceneId + 1) + u" (" + sceneName + u") .")
+        self.debugLog("saveScene - finished Validation for Scene " + str(sceneId + 1) + " (" + sceneName + ") .")
 
         colorDict = dict()
         if valuesDict['setting1Value']:
@@ -680,7 +680,7 @@ class Plugin(indigo.PluginBase):
         if len(circDict) > 0:
             newSceneDict['circadianData'] = circDict
 
-        self.debugLog(u"saveScene - adding newSceneDict=" + str(newSceneDict))
+        self.debugLog("saveScene - adding newSceneDict=" + str(newSceneDict))
 
         scenesList = self.pluginPrefs['scenesList']
 
@@ -691,9 +691,9 @@ class Plugin(indigo.PluginBase):
             aDict['sceneName'] = scenesList[aNumber]['sceneName']
             aDict['sceneType'] = scenesList[aNumber]['sceneType']
             aDict['sceneInterval'] = scenesList[aNumber]['sceneInterval']
-            if 'colorData' in scenesList[aNumber].keys():
+            if 'colorData' in list(scenesList[aNumber].keys()):
                 aDict['colorData'] = scenesList[aNumber]['colorData']
-            if 'circadianData' in scenesList[aNumber].keys():
+            if 'circadianData' in list(scenesList[aNumber].keys()):
                 aDict['circadianData'] = scenesList[aNumber]['circadianData']
 
             if aNumber == sceneId:
@@ -705,12 +705,12 @@ class Plugin(indigo.PluginBase):
                 if len(circDict) > 0:
                     aDict['circadianData'] = circDict
 
-                self.debugLog(u"saveScene - finished adding NEW, sceneDict=" + str(aDict))
+                self.debugLog("saveScene - finished adding NEW, sceneDict=" + str(aDict))
 
         # Save the device's states to the preset.
         self.pluginPrefs['scenesList'] = newScenesList
 
-        indigo.server.log(u"saveScene - states saved to Scene " + unicode(sceneId + 1) + u" (" + sceneName + u")")
+        indigo.server.log("saveScene - states saved to Scene " + str(sceneId + 1) + " (" + sceneName + ")")
 
 
     def getMenuActionConfigUiValues(self, menuId):
@@ -719,7 +719,7 @@ class Plugin(indigo.PluginBase):
 
         if menuId == "saveScene":
             scenesList = self.pluginPrefs.get('scenesList', None)
-            self.debugLog(u"getMenuActionConfigUiValues: get initial value for saveScene dialog.")
+            self.debugLog("getMenuActionConfigUiValues: get initial value for saveScene dialog.")
 
             theScene = None
 
@@ -728,49 +728,49 @@ class Plugin(indigo.PluginBase):
 
                 for sceneDict in scenesList:
                     # Determine whether the Scene has saved data or not.
-                    if 'colorData' in sceneDict.keys() or 'circadianData' in sceneDict.keys():
+                    if 'colorData' in list(sceneDict.keys()) or 'circadianData' in list(sceneDict.keys()):
                         if sceneNumber is 0:
                             theScene = sceneDict
                     sceneNumber += 1
 
             if theScene is not None:
-                self.debugLog(u"...getMenuActionConfigUiValues: loading data for scene0:\n" + unicode(theScene))
+                self.debugLog("...getMenuActionConfigUiValues: loading data for scene0:\n" + str(theScene))
 
                 valuesDict['sceneName'] = theScene['sceneName']
                 valuesDict['sceneType'] = theScene['sceneType']
                 valuesDict['sceneInterval'] = theScene['sceneInterval']
-                if 'colorData' in theScene.keys():
+                if 'colorData' in list(theScene.keys()):
                     colorData = theScene['colorData']
                     if colorData is not None:
-                        if 'setting1Value' in colorData.keys():
+                        if 'setting1Value' in list(colorData.keys()):
                             valuesDict['setting1Value'] = colorData['setting1Value']
                         else:
                             valuesDict['setting1Value'] = ""
-                        if 'setting2Value' in colorData.keys():
+                        if 'setting2Value' in list(colorData.keys()):
                             valuesDict['setting2Value'] = colorData['setting2Value']
                         else:
                             valuesDict['setting2Value'] = ""
-                        if 'setting3Value' in colorData.keys():
+                        if 'setting3Value' in list(colorData.keys()):
                             valuesDict['setting3Value'] = colorData['setting3Value']
                         else:
                             valuesDict['setting3Value'] = ""
-                        if 'setting4Value' in colorData.keys():
+                        if 'setting4Value' in list(colorData.keys()):
                             valuesDict['setting4Value'] = colorData['setting4Value']
                         else:
                             valuesDict['setting4Value'] = ""
-                        if 'setting5Value' in colorData.keys():
+                        if 'setting5Value' in list(colorData.keys()):
                             valuesDict['setting5Value'] = colorData['setting5Value']
                         else:
                             valuesDict['setting5Value'] = ""
 
-                if 'circadianData' in theScene.keys():
+                if 'circadianData' in list(theScene.keys()):
                     circData = theScene['circadianData']
                     if circData is not None:
-                        if 'CircadianColorTempValues' in circData.keys():
+                        if 'CircadianColorTempValues' in list(circData.keys()):
                             valuesDict['CircadianColorTempValues'] = circData['CircadianColorTempValues']
                         else:
                             valuesDict['CircadianColorTempValues'] = ""
-                        if 'CircadianBrightnessValues' in circData.keys():
+                        if 'CircadianBrightnessValues' in list(circData.keys()):
                             valuesDict['CircadianBrightnessValues'] = circData['CircadianBrightnessValues']
                         else:
                             valuesDict['CircadianBrightnessValues'] = ""
@@ -792,9 +792,9 @@ class Plugin(indigo.PluginBase):
             sceneName = sceneDict['sceneName']
             sceneType = sceneDict['sceneType']
             sceneInterval = int(sceneDict['sceneInterval'])
-            if 'colorData' in sceneDict.keys():
+            if 'colorData' in list(sceneDict.keys()):
                 colorData = sceneDict['colorData']
-            if 'circadianData' in sceneDict.keys():
+            if 'circadianData' in list(sceneDict.keys()):
                 circadianData = sceneDict['circadianData']
 
             indigo.server.log("startScene LightifyGroup: " + str(indigoDevice.name) + ", scene: " + str(sceneName) +
@@ -804,7 +804,7 @@ class Plugin(indigo.PluginBase):
             if sceneType != "circadian":
                 for settingNum in range(1,6):
                     key = 'setting' + str(settingNum) + 'Value'
-                    if key in colorData.keys():
+                    if key in list(colorData.keys()):
                         thesetting = colorData[key]
                         self.debugLog("...settingNum=" + str(settingNum) + ", data=" + str(thesetting))
                         tempArray = thesetting.split(",")
@@ -975,18 +975,18 @@ class Plugin(indigo.PluginBase):
     def validateRGBWConfig(self, configStr, valuesDict, errorsDict):
 
         if not valuesDict[configStr]:
-            errorsDict[configStr] =  u"No " + configStr + u" specified."
+            errorsDict[configStr] =  "No " + configStr + " specified."
             errorsDict['showAlertText'] += errorsDict[configStr]
-            errorText = u"No " + configStr + u" specified."
+            errorText = "No " + configStr + " specified."
             self.errorLog(errorText)
             #return valuesDict, errorsDict
             return False, valuesDict, errorsDict
 
         tempArray = valuesDict[configStr].split(",")
         if len(tempArray) != 6:
-            errorsDict[configStr] =  configStr + u" must have 6 values. R,G,B,Temp,Brightness,TransMilli"
+            errorsDict[configStr] =  configStr + " must have 6 values. R,G,B,Temp,Brightness,TransMilli"
             errorsDict['showAlertText'] += errorsDict[configStr]
-            errorText = configStr + u" must have 6 values. R,G,B,Temp,Brightness,TransMilli"
+            errorText = configStr + " must have 6 values. R,G,B,Temp,Brightness,TransMilli"
             self.errorLog(errorText)
             return False, valuesDict, errorsDict
 
@@ -1003,37 +1003,37 @@ class Plugin(indigo.PluginBase):
             isTemp = True
 
         if isTemp == False and (redVal < 0 or redVal > 255 or greenVal < 0 or greenVal > 255 or blueVal < 0 or blueVal > 255):
-            errorsDict[configStr] =  configStr + u" RGB values must be between 0 and 255."
+            errorsDict[configStr] =  configStr + " RGB values must be between 0 and 255."
             errorsDict['showAlertText'] += errorsDict[configStr]
-            errorText = configStr + u" RGB values must be between 0 and 255."
+            errorText = configStr + " RGB values must be between 0 and 255."
             self.errorLog(errorText)
             return False, valuesDict, errorsDict
 
         if isTemp == True and (redVal != 255 or greenVal != 255 or blueVal != 255):
-            errorsDict[configStr] =  configStr + u" ColorTemp RGB values must be 255."
+            errorsDict[configStr] =  configStr + " ColorTemp RGB values must be 255."
             errorsDict['showAlertText'] += errorsDict[configStr]
-            errorText = configStr + u" ColorTemp RGB values must be 255."
+            errorText = configStr + " ColorTemp RGB values must be 255."
             self.errorLog(errorText)
             return False, valuesDict, errorsDict
 
         if isTemp == True and (colorTemp < 1900 or colorTemp > 6500):
-            errorsDict[configStr] =  configStr + u" ColorTemp must be between 1900 and 2500."
+            errorsDict[configStr] =  configStr + " ColorTemp must be between 1900 and 2500."
             errorsDict['showAlertText'] += errorsDict[configStr]
-            errorText = configStr + u" ColorTemp must be between 1900 and 2500."
+            errorText = configStr + " ColorTemp must be between 1900 and 2500."
             self.errorLog(errorText)
             return False, valuesDict, errorsDict
 
         if brightness < 1 or brightness > 100:
-            errorsDict[configStr] =  configStr + u" Brightness must be between 1 and 100."
+            errorsDict[configStr] =  configStr + " Brightness must be between 1 and 100."
             errorsDict['showAlertText'] += errorsDict[configStr]
-            errorText = configStr + u" Brightness must be between 1 and 100."
+            errorText = configStr + " Brightness must be between 1 and 100."
             self.errorLog(errorText)
             return False, valuesDict, errorsDict
 
         if transMilli < 10 or transMilli > 100000:
-            errorsDict[configStr] =  configStr + u" transMilli must be between 10 and 100000."
+            errorsDict[configStr] =  configStr + " transMilli must be between 10 and 100000."
             errorsDict['showAlertText'] += errorsDict[configStr]
-            errorText = configStr + u" transMilli must be between 10 and 100000."
+            errorText = configStr + " transMilli must be between 10 and 100000."
             self.errorLog(errorText)
             return False, valuesDict, errorsDict
 
@@ -1050,18 +1050,18 @@ class Plugin(indigo.PluginBase):
         # Late Night&lt;Pre-Sunrise&lt;Post-Sunrise&lt;AM Peak&lt;Max, also Max&gt;Pre-Sunset&gt;Post-Sunset&gt;Late Night
 
         if not valuesDict[configStr]:
-            errorsDict[configStr] =  u"No " + configStr + u" specified."
+            errorsDict[configStr] =  "No " + configStr + " specified."
             errorsDict['showAlertText'] += errorsDict[configStr]
-            errorText = u"No " + configStr + u" specified."
+            errorText = "No " + configStr + " specified."
             self.errorLog(errorText)
             #return valuesDict, errorsDict
             return False, valuesDict, errorsDict
 
         tempArray = valuesDict[configStr].split(",")
         if len(tempArray) != 7:
-            errorsDict[configStr] =  configStr + u" must have 7 values. Late Night,Pre-Sunrise,Post-Sunrise,AM Peak,Max Temp,Pre-Sunset,Post-Sunset."
+            errorsDict[configStr] =  configStr + " must have 7 values. Late Night,Pre-Sunrise,Post-Sunrise,AM Peak,Max Temp,Pre-Sunset,Post-Sunset."
             errorsDict['showAlertText'] += errorsDict[configStr]
-            errorText = configStr + u" must have 7 values. Late Night,Pre-Sunrise,Post-Sunrise,AM Peak,Max Temp,Pre-Sunset,Post-Sunset."
+            errorText = configStr + " must have 7 values. Late Night,Pre-Sunrise,Post-Sunrise,AM Peak,Max Temp,Pre-Sunset,Post-Sunset."
             self.errorLog(errorText)
             return False, valuesDict, errorsDict
 
@@ -1075,16 +1075,16 @@ class Plugin(indigo.PluginBase):
         postSunset = int(tempArray[6])
 
         if lateNight > preSunrise or preSunrise > postSunrise or postSunrise > amPeak or amPeak > maxVal:
-            errorsDict[configStr] =  configStr + u" - invalid values: Ensure Late Night > Pre-Sunrise > Post-Sunrise > AM Peak > Max."
+            errorsDict[configStr] =  configStr + " - invalid values: Ensure Late Night > Pre-Sunrise > Post-Sunrise > AM Peak > Max."
             errorsDict['showAlertText'] += errorsDict[configStr]
-            errorText = configStr + u" - invalid values: Ensure Late Night > Pre-Sunrise > Post-Sunrise > AM Peak > Max."
+            errorText = configStr + " - invalid values: Ensure Late Night > Pre-Sunrise > Post-Sunrise > AM Peak > Max."
             self.errorLog(errorText)
             return False, valuesDict, errorsDict
 
         if maxVal < preSunset or preSunset < postSunset or postSunset < lateNight:
-            errorsDict[configStr] =  configStr + u" Invalid values: Ensure Max > Pre-Sunset > Post-Sunset > Late Night."
+            errorsDict[configStr] =  configStr + " Invalid values: Ensure Max > Pre-Sunset > Post-Sunset > Late Night."
             errorsDict['showAlertText'] += errorsDict[configStr]
-            errorText = configStr + u" - invalid values: Ensure Max > Pre-Sunset > Post-Sunset > Late Night."
+            errorText = configStr + " - invalid values: Ensure Max > Pre-Sunset > Post-Sunset > Late Night."
             self.errorLog(errorText)
             return False, valuesDict, errorsDict
 
@@ -1092,9 +1092,9 @@ class Plugin(indigo.PluginBase):
                 or postSunrise < 1900 or postSunrise > 6500 or amPeak < 1900 or amPeak > 6500 \
                 or maxVal < 1900 or maxVal > 6500 or preSunset < 1900 or preSunset > 6500 \
                 or postSunset < 1900 or postSunset > 6500):
-            errorsDict[configStr] =  configStr + u" values must be between 1900 and 6500."
+            errorsDict[configStr] =  configStr + " values must be between 1900 and 6500."
             errorsDict['showAlertText'] += errorsDict[configStr]
-            errorText = configStr + u" values must be between 1900 and 6500."
+            errorText = configStr + " values must be between 1900 and 6500."
             self.errorLog(errorText)
             return False, valuesDict, errorsDict
 
@@ -1102,9 +1102,9 @@ class Plugin(indigo.PluginBase):
                or postSunrise < 1 or postSunrise > 100 or amPeak < 1 or amPeak > 100 \
                or maxVal < 1 or maxVal > 100 or preSunset < 1 or preSunset > 100 \
                or postSunset < 0 or postSunset > 100):
-            errorsDict[configStr] =  configStr + u" values must be between 1 and 100."
+            errorsDict[configStr] =  configStr + " values must be between 1 and 100."
             errorsDict['showAlertText'] += errorsDict[configStr]
-            errorText = configStr + u" values must be between 1 and 100."
+            errorText = configStr + " values must be between 1 and 100."
             self.errorLog(errorText)
             return False, valuesDict, errorsDict
 
@@ -1140,10 +1140,10 @@ class Plugin(indigo.PluginBase):
         redLevel = "0"
         greenLevel = "0"
         blueLevel = "0"
-        if indigoDevice.supportsWhiteTemperature is True and 'whiteTemperature.ui' in indigoDevice.states.keys():
+        if indigoDevice.supportsWhiteTemperature is True and 'whiteTemperature.ui' in list(indigoDevice.states.keys()):
             whiteTemp = indigoDevice.states['whiteTemperature.ui']
-        if indigoDevice.supportsRGB is True and 'redLevel.ui' in indigoDevice.states.keys() \
-                and 'greenLevel.ui' in indigoDevice.states.keys() and 'blueLevel.ui' in indigoDevice.states.keys():
+        if indigoDevice.supportsRGB is True and 'redLevel.ui' in list(indigoDevice.states.keys()) \
+                and 'greenLevel.ui' in list(indigoDevice.states.keys()) and 'blueLevel.ui' in list(indigoDevice.states.keys()):
             redLevel = indigoDevice.states['redLevel.ui']
             greenLevel = indigoDevice.states['greenLevel.ui']
             blueLevel = indigoDevice.states['blueLevel.ui']
@@ -1185,18 +1185,18 @@ class Plugin(indigo.PluginBase):
     # Methods for the dimmer device
     ########################################
     def actionControlDevice(self, action, dev, gotLock=False):
-        self.debugLog(u"...actionControlDevice - device=\"%s\", action=%s" % (dev.name, action.deviceAction))
+        self.debugLog("...actionControlDevice - device=\"%s\", action=%s" % (dev.name, action.deviceAction))
         ###### First get the lightify device group
         theGroup = self.getLightifyGroup(dev)
         if theGroup is None:
             sendSuccess = False  # Set to False if it failed.
-            self.debugLog(u"...actionControlDevice - lightify group not found \"%s\"" % (dev.name), isError=True)
+            self.debugLog("...actionControlDevice - lightify group not found \"%s\"" % (dev.name), isError=True)
         else:
             try:
                 if gotLock is not True:
                     self.lightifyLock.acquire()
 
-                self.debugLog(u"...actionControlDevice - lightify group found \"%s\"" % (dev.name))
+                self.debugLog("...actionControlDevice - lightify group found \"%s\"" % (dev.name))
                 ###### TURN ON ######
                 if action.deviceAction == indigo.kDeviceAction.TurnOn:
                     # Command hardware module (dev) to turn ON here:
@@ -1205,20 +1205,20 @@ class Plugin(indigo.PluginBase):
 
                     if sendSuccess:
                         # If success then log that the command was successfully sent.
-                        indigo.server.log(u"\"%s\" %s" % (dev.name, "on"))
+                        indigo.server.log("\"%s\" %s" % (dev.name, "on"))
 
                         # And then tell the Indigo Server to update the state.
                         #dev.updateStateOnServer("onOffState", True)
                         dev.updateStateOnServer('onOffState', value=True, uiValue='on')
                     else:
                         # Else log failure but do NOT update state on Indigo Server.
-                        indigo.server.log(u"\"%s\" %s failed" % (dev.name, "on"), isError=True)
+                        indigo.server.log("\"%s\" %s failed" % (dev.name, "on"), isError=True)
 
                 ###### TURN OFF ######
                 elif action.deviceAction == indigo.kDeviceAction.TurnOff:
 
                     activeScene = dev.states.get('activeScene', None)
-                    self.debugLog(u"...actionControlDevice - activeScene= \"%s\"" % (activeScene))
+                    self.debugLog("...actionControlDevice - activeScene= \"%s\"" % (activeScene))
 
                     if activeScene is not None:
                         self.stopSceneThread(dev)
@@ -1232,14 +1232,14 @@ class Plugin(indigo.PluginBase):
 
                     if sendSuccess:
                         # If success then log that the command was successfully sent.
-                        indigo.server.log(u"\"%s\" %s" % (dev.name, "off"))
+                        indigo.server.log("\"%s\" %s" % (dev.name, "off"))
 
                         # And then tell the Indigo Server to update the state:
                         dev.updateStateOnServer('onOffState', value=False, uiValue='off')
                         dev.updateStateOnServer("brightnessLevel", 0)
                     else:
                         # Else log failure but do NOT update state on Indigo Server.
-                        indigo.server.log(u"\"%s\" %s failed" % (dev.name, "off"), isError=True)
+                        indigo.server.log("\"%s\" %s failed" % (dev.name, "off"), isError=True)
 
                 ###### TOGGLE ######
                 elif action.deviceAction == indigo.kDeviceAction.Toggle:
@@ -1250,13 +1250,13 @@ class Plugin(indigo.PluginBase):
 
                     if sendSuccess:
                         # If success then log that the command was successfully sent.
-                        indigo.server.log(u"\"%s\" %s" % (dev.name, "toggle"))
+                        indigo.server.log("\"%s\" %s" % (dev.name, "toggle"))
 
                         # And then tell the Indigo Server to update the state:
                         dev.updateStateOnServer("onOffState", newOnState)
                     else:
                         # Else log failure but do NOT update state on Indigo Server.
-                        indigo.server.log(u"\"%s\" %s failed" % (dev.name, "toggle"), isError=True)
+                        indigo.server.log("\"%s\" %s failed" % (dev.name, "toggle"), isError=True)
 
                 ###### SET BRIGHTNESS ######
                 elif action.deviceAction == indigo.kDeviceAction.SetBrightness:
@@ -1270,13 +1270,13 @@ class Plugin(indigo.PluginBase):
 
                     if sendSuccess:
                         # If success then log that the command was successfully sent.
-                        indigo.server.log(u"\"%s\" %s to %d" % (dev.name, "set brightness", newBrightness))
+                        indigo.server.log("\"%s\" %s to %d" % (dev.name, "set brightness", newBrightness))
 
                         # And then tell the Indigo Server to update the state:
                         dev.updateStateOnServer("brightnessLevel", newBrightness)
                     else:
                         # Else log failure but do NOT update state on Indigo Server.
-                        indigo.server.log(u"\"%s\" %s to %d failed" % (dev.name, "set brightness", newBrightness),
+                        indigo.server.log("\"%s\" %s to %d failed" % (dev.name, "set brightness", newBrightness),
                                           isError=True)
 
                 ###### BRIGHTEN BY ######
@@ -1290,13 +1290,13 @@ class Plugin(indigo.PluginBase):
 
                     if sendSuccess:
                         # If success then log that the command was successfully sent.
-                        indigo.server.log(u"\"%s\" %s to %d" % (dev.name, "brighten", newBrightness))
+                        indigo.server.log("\"%s\" %s to %d" % (dev.name, "brighten", newBrightness))
 
                         # And then tell the Indigo Server to update the state:
                         dev.updateStateOnServer("brightnessLevel", newBrightness)
                     else:
                         # Else log failure but do NOT update state on Indigo Server.
-                        indigo.server.log(u"\"%s\" %s to %d failed" % (dev.name, "brighten", newBrightness),
+                        indigo.server.log("\"%s\" %s to %d failed" % (dev.name, "brighten", newBrightness),
                                           isError=True)
 
                 ###### DIM BY ######
@@ -1310,13 +1310,13 @@ class Plugin(indigo.PluginBase):
 
                     if sendSuccess:
                         # If success then log that the command was successfully sent.
-                        indigo.server.log(u"\"%s\" %s to %d" % (dev.name, "dim", newBrightness))
+                        indigo.server.log("\"%s\" %s to %d" % (dev.name, "dim", newBrightness))
 
                         # And then tell the Indigo Server to update the state:
                         dev.updateStateOnServer("brightnessLevel", newBrightness)
                     else:
                         # Else log failure but do NOT update state on Indigo Server.
-                        indigo.server.log(u"\"%s\" %s to %d failed" % (dev.name, "dim", newBrightness),
+                        indigo.server.log("\"%s\" %s to %d failed" % (dev.name, "dim", newBrightness),
                                           isError=True)
 
                 ###### SET COLOR LEVELS ######
@@ -1363,7 +1363,7 @@ class Plugin(indigo.PluginBase):
                             brightnessByte = int(round(255.0 * (brightness / 100.0)))
 
                             if dev.supportsRGB == False and channel == 'whiteTemperature' and brightness > 0 and brightness < 2700:
-                                indigo.server.log(u"\"%s\" is tunable-white - cannot set lower than 2700K (%s) " % (dev.name, actionColorVals[channel]))
+                                indigo.server.log("\"%s\" is tunable-white - cannot set lower than 2700K (%s) " % (dev.name, actionColorVals[channel]))
                                 # Lightify Tunable White bulbs cannot go lower than 2700K
                                 brightness = float(2700)
 
@@ -1406,13 +1406,13 @@ class Plugin(indigo.PluginBase):
                     # def set_rgb(self, r, g, b, time):
                     if temperature > 0 and red == 255 and green == 255 and blue == 255:
                         self.debugLog(
-                            u"...SetColor lightify set_temperature - for \"%s\" red=%s, green=%s, blue=%s, temp=%s" % (
+                            "...SetColor lightify set_temperature - for \"%s\" red=%s, green=%s, blue=%s, temp=%s" % (
                                 dev.name, str(red), str(green), str(blue), str(temperature)))
                         theGroup.set_temperature(temperature, 50)
                     if red != 255 or green != 255 or blue != 255:
                         if temperature == 0:
                             self.debugLog(
-                                u"...SetColor lightify set_rgb - for \"%s\" red=%s, green=%s, blue=%s, temp=%s" % (
+                                "...SetColor lightify set_rgb - for \"%s\" red=%s, green=%s, blue=%s, temp=%s" % (
                                     dev.name, str(red), str(green), str(blue), str(temperature)))
                             theGroup.set_rgb(red, green, blue, 50)
                     # Set to False if it failed.
@@ -1421,17 +1421,17 @@ class Plugin(indigo.PluginBase):
                     resultValsStr = ' '.join(resultVals)
                     if sendSuccess:
                         # If success then log that the command was successfully sent.
-                        indigo.server.log(u"\"%s\" %s to %s" % (dev.name, "set color", resultValsStr))
+                        indigo.server.log("\"%s\" %s to %s" % (dev.name, "set color", resultValsStr))
 
                         # And then tell the Indigo Server to update the color level states:
                         if len(keyValueList) > 0:
                             dev.updateStatesOnServer(keyValueList)
                     else:
                         # Else log failure but do NOT update state on Indigo Server.
-                        indigo.server.log(u"\"%s\" %s to %s failed" % (dev.name, "set color", resultValsStr),
+                        indigo.server.log("\"%s\" %s to %s failed" % (dev.name, "set color", resultValsStr),
                                           isError=True)
                 else:
-                    indigo.server.log(u"Unknown action for " + dev.name + ", action=" + action.deviceAction)
+                    indigo.server.log("Unknown action for " + dev.name + ", action=" + action.deviceAction)
 
                 self.updateUIForScene(dev)
 
@@ -1898,7 +1898,7 @@ class IndigoLogHandler(logging.Handler, object):
         # For any level besides INFO and ERROR (which Indigo handles), we append
         # the debug level (i.e. Critical, Warning, Debug, etc) to the type string
         if record.levelno not in (logging.INFO, logging.ERROR):
-            type_string += u" %s" % record.levelname.title()
+            type_string += " %s" % record.levelname.title()
         # Then we write the message
         indigo.server.log(message=self.format(record), type=type_string, isError=is_error)
